requirements.txt
pandas==1.4.2
python-dotenv==0.20.0
requests==2.25.1